Intellij support

+++ src/fr/simplgame/pss/event/BotListener.java
@@ -3,11 +3,14 @@ package fr.simplgame.pss.event;
 import java.awt.Color;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
+import fr.simplgame.pss.PSS;
 import fr.simplgame.pss.command.CommandMap;
 import fr.simplgame.pss.util.CSV;
 import fr.simplgame.pss.util.Loader;
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.ChannelType;
 import net.dv8tion.jda.api.events.GenericEvent;
 import net.dv8tion.jda.api.events.guild.member.GuildMemberJoinEvent;
@@ -15,6 +18,7 @@ import net.dv8tion.jda.api.events.guild.member.GuildMemberLeaveEvent;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.exceptions.HierarchyException;
 import net.dv8tion.jda.api.hooks.EventListener;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * Permet l'écoute de chaque event
@@ -47,7 +51,7 @@ public class BotListener implements EventListener {
 	}
 
 	@Override
-	public void onEvent(GenericEvent event) {
+	public void onEvent(@NotNull GenericEvent event) {
 		if (event instanceof MessageReceivedEvent)
 			onMessage((MessageReceivedEvent) event);
 		if (event instanceof GuildMemberJoinEvent)
@@ -63,14 +67,13 @@ public class BotListener implements EventListener {
 					"./res/server.csv");
 
 		String joinId = CSV.getCell(gmje.getGuild().getId(), "join_channel", "./res/server.csv");
-		if (!joinId.equals("NaN") || !joinId.equals(" "))
-			gmje.getGuild().getTextChannelById(joinId)
-					.sendMessage(CSV.getCell(gmje.getGuild().getId(), "join_message", "./res/server.csv")
-							.replace("[USER]", gmje.getMember().getEffectiveName()))
-					.queue();
+		Objects.requireNonNull(gmje.getGuild().getTextChannelById(joinId))
+				.sendMessage(CSV.getCell(gmje.getGuild().getId(), "join_message", "./res/server.csv")
+						.replace("[USER]", gmje.getMember().getEffectiveName()).replace("[COMMA_DOT]", ";"))
+				.queue();
 
 		String roleId = CSV.getCell(gmje.getGuild().getId(), "join_role", "./res/server.csv");
-		String roleWrong = "NaN";
+		StringBuilder roleWrong = new StringBuilder("NaN");
 		Loader loader = null;
 		boolean roleError = false;
 		for (Loader l : loaders) {
@@ -79,32 +82,32 @@ public class BotListener implements EventListener {
 				break;
 			}
 		}
-		if (!roleId.equals("NaN") || !roleId.equals(" ")) {
-			String[] roles = roleId.split("_");
-			for (String role : roles) {
-				try {
-					gmje.getGuild().addRoleToMember(gmje.getMember().getId(), gmje.getGuild().getRoleById(role))
-							.queue();
-				} catch (HierarchyException e) {
-					if (roleWrong.equals("NaN"))
-						roleWrong = gmje.getGuild().getRoleById(role).getAsMention();
-					else
-						roleWrong += loader.lang.get("general.mark.comma")
-								+ gmje.getGuild().getRoleById(role).getAsMention();
-					roleError = true;
+		String[] roles = roleId.split("_");
+		for (String role : roles) {
+			try {
+				gmje.getGuild().addRoleToMember(gmje.getMember().getId(), Objects.requireNonNull(gmje.getGuild().getRoleById(role)))
+						.queue();
+			} catch (HierarchyException e) {
+				if (roleWrong.toString().equals("NaN"))
+					roleWrong = new StringBuilder(Objects.requireNonNull(gmje.getGuild().getRoleById(role)).getAsMention());
+				else {
+					assert loader != null;
+					roleWrong.append(loader.lang.get("general.mark.comma")).append(Objects.requireNonNull(gmje.getGuild().getRoleById(role)).getAsMention());
 				}
+				roleError = true;
 			}
 		}
 
-		roleWrong += loader.lang.get("general.mark.dot");
+		assert loader != null;
+		roleWrong.append(loader.lang.get("general.mark.dot"));
 		if (roleError) {
 			if (!CSV.getCell(gmje.getGuild().getId(), "log_channel", "./res/server.csv").equals("NaN")) {
 				EmbedBuilder embed = new EmbedBuilder();
 				embed.setTitle(loader.lang.get("server.error.jls.L1"));
 				embed.setDescription(loader.lang.get("server.error.jls.L2").replace("\\n[ROLE]", "\n" + roleWrong));
 				embed.setColor(Color.decode("#D61C1C"));
-				gmje.getGuild()
-						.getTextChannelById(CSV.getCell(gmje.getGuild().getId(), "log_channel", "./res/server.csv"))
+				Objects.requireNonNull(gmje.getGuild()
+						.getTextChannelById(CSV.getCell(gmje.getGuild().getId(), "log_channel", "./res/server.csv")))
 						.sendMessage(embed.build()).queue();
 			}
 		}
@@ -116,11 +119,10 @@ public class BotListener implements EventListener {
 					"./res/server.csv");
 
 		String leaveId = CSV.getCell(gmle.getGuild().getId(), "leave_channel", "./res/server.csv");
-		if (!leaveId.equals("NaN") || !leaveId.equals(" "))
-			gmle.getGuild().getTextChannelById(leaveId)
-					.sendMessage(CSV.getCell(gmle.getGuild().getId(), "leave_message", "./res/server.csv")
-							.replace("[USER]", gmle.getMember().getEffectiveName()))
-					.queue();
+		Objects.requireNonNull(gmle.getGuild().getTextChannelById(leaveId))
+				.sendMessage(CSV.getCell(gmle.getGuild().getId(), "leave_message", "./res/server.csv")
+						.replace("[USER]", gmle.getMember().getEffectiveName()).replace("[COMMA_DOT]", ";"))
+				.queue();
 	}
 
 	private void onMessage(MessageReceivedEvent mre) {
@@ -138,6 +140,9 @@ public class BotListener implements EventListener {
 			lang[1] = new Loader();
 		else
 			lang[1] = new Loader(serverLanguage);
+		
+		System.out.println(userLanguage);
+		System.out.println(serverLanguage);
 
 		// If the user is PSS, returning to the top function.
 		if (mre.getAuthor().equals(mre.getJDA().getSelfUser()))
@@ -148,10 +153,11 @@ public class BotListener implements EventListener {
 			if (mre.getAuthor().isBot())
 				return;
 			message = message.replaceFirst(commandMap.getTag(), "");
-			if (commandMap.commandUser(mre.getAuthor(), message, mre.getMessage(), lang)) {
+			if (commandMap.commandUser(mre.getAuthor(), message, mre.getMessage(), lang))
 				if (mre.getChannel().getType() == ChannelType.TEXT)
-					mre.getMessage().delete().queue();
-			}
+					if (Objects.requireNonNull(mre.getGuild().getMemberById(PSS.jda.getSelfUser().getId()))
+							.hasPermission(Permission.MESSAGE_MANAGE))
+						mre.getMessage().delete().queue();
 		}
 	}
